added support for cpp tests to Moai test framework

+++ test/main-tests.lua
@@ -1,47 +1,59 @@
-stagingDir = MOAIFileSystem.getAbsoluteDirectoryPath ( 'staging' )
-testingDir = MOAIFileSystem.getAbsoluteDirectoryPath ( 'testing' )
+dofile ( 'util.lua' )
+
 rootDir = MOAIFileSystem.getWorkingDirectory ()
-
 results = rootDir .. '.results'
-allResults = rootDir .. 'test-results'
-
-MOAIFileSystem.affirmPath ( stagingDir )
-MOAIFileSystem.affirmPath ( testingDir )
-
-stagingList = MOAIFileSystem.listDirectories ( stagingDir )
-
-command = string.format ( '%%MOAI_BIN%%\\moai-test -r "%s" main.lua', results )
-
+allResults = rootDir .. 'testing-results'
 MOAIFileSystem.deleteFile ( allResults )
 
-for i, dirname in ipairs ( stagingList ) do
-	
-	local from = stagingDir .. dirname
-	local to = testingDir .. dirname
+-- run the C++ tests
+stagingDir = MOAIFileSystem.getAbsoluteDirectoryPath ( 'staging-cpp' )
+if MOAIFileSystem.checkPathExists ( stagingDir ) == true then
 
-	if MOAIFileSystem.checkPathExists ( to ) == true then
-		MOAIFileSystem.deleteDirectory ( to )
+	testingDir = MOAIFileSystem.getAbsoluteDirectoryPath ( 'testing-cpp' )
+	MOAIFileSystem.affirmPath ( testingDir )
+
+	local testList = MOAITestMgr.getTestList ()
+	for i, testname in ipairs ( testList ) do
+		
+		local from = stagingDir .. testname
+		local to = testingDir .. testname
+		
+		if MOAIFileSystem.checkPathExists ( to ) == true then
+			MOAIFileSystem.deleteDirectory ( to )
+		end
+		
+		if MOAIFileSystem.checkPathExists ( from ) == true then
+			MOAIFileSystem.copy ( from, to )
+			MOAIFileSystem.setWorkingDirectory ( to )
+			os.execute ( string.format ( '%%MOAI_BIN%%\\moai-test -r "%s" -t "%s"', results, testname ))
+			MOAIFileSystem.setWorkingDirectory ( rootDir )
+			gatherResults ( results, allResults )
+		end
 	end
-	
-	MOAIFileSystem.copy ( from, to )
-	
-	MOAIFileSystem.setWorkingDirectory ( to )
-	os.execute ( command )
-	MOAIFileSystem.setWorkingDirectory ( rootDir )
-	
-	if MOAIFileSystem.checkFileExists ( results ) then
+end
+
+-- run the Lua tests
+stagingDir = MOAIFileSystem.getAbsoluteDirectoryPath ( 'staging-lua' )
+if MOAIFileSystem.checkPathExists ( stagingDir ) == true then
+
+	testingDir = MOAIFileSystem.getAbsoluteDirectoryPath ( 'testing-lua' )
+	MOAIFileSystem.affirmPath ( testingDir )
+
+	stagingList = MOAIFileSystem.listDirectories ( stagingDir )
+	for i, dirname in ipairs ( stagingList ) do
 		
-		inFile = MOAIFileStream.new ()
-		outFile = MOAIFileStream.new ()
+		local from = stagingDir .. dirname
+		local to = testingDir .. dirname
+
+		if MOAIFileSystem.checkPathExists ( to ) == true then
+			MOAIFileSystem.deleteDirectory ( to )
+		end
 		
-		inFile:open ( results, MOAIFileStream.READ )
-		outFile:open ( allResults, MOAIFileStream.READ_WRITE_AFFIRM )
-		
-		outFile:seek ( 0, MOAIFileStream.SEEK_END )
-		outFile:writeStream ( inFile )
-		
-		inFile:close ()
-		outFile:close ()
+		MOAIFileSystem.copy ( from, to )
+		MOAIFileSystem.setWorkingDirectory ( to )
+		os.execute ( string.format ( '%%MOAI_BIN%%\\moai-test -r "%s" main.lua', results ))
+		MOAIFileSystem.setWorkingDirectory ( rootDir )
+		gatherResults ( results, allResults )
 	end
 end
